hello
this is updated and will keep a constant angular velocity when rotating. However, the issue with this is that its translational speed when doing so will not remain constant. I have not yet found a way around this. To do so, I will need to come together with a few people and derive a mathematical algorithm for each wheel.